Fix the broken migrations
