+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-1.json
@@ -4,22 +4,41 @@
         "name": "core/gallery",
         "isValid": true,
         "attributes": {
-            "images": [
-                {
-                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
-                    "alt": "title"
-                },
-                {
-                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
-                    "alt": "title"
-                }
-            ],
             "columns": 2,
             "imageCrop": true,
             "linkTo": "none",
-            "align": "wide"
+            "imageCount": 2,
+            "allowResize": false,
+            "isListItem": true
         },
-        "innerBlocks": [],
+        "innerBlocks": [
+            {
+                "clientId": "_clientId_0",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
+                    "alt": "title",
+                    "id": null,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            },
+            {
+                "clientId": "_clientId_1",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
+                    "alt": "title",
+                    "id": null,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            }
+        ],
         "originalContent": "<div class=\"wp-block-gallery columns-2 is-cropped alignwide\">\n\t<figure class=\"blocks-gallery-image\">\n\t\t<img src=\"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==\" alt=\"title\" />\n\t</figure>\n\t<figure class=\"blocks-gallery-image\">\n\t\t<img src=\"data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=\" alt=\"title\" />\n\t</figure>\n</div>"
     }
 ]
