+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-2.json
@@ -4,27 +4,43 @@
         "name": "core/gallery",
         "isValid": true,
         "attributes": {
-            "images": [
-                {
-                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
-                    "alt": "title",
-                    "id": "1",
-                    "caption": []
-                },
-                {
-                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
-                    "alt": "title",
-                    "id": "2",
-                    "caption": []
-                }
-            ],
-            "ids": [],
             "columns": 2,
             "imageCrop": true,
             "linkTo": "none",
-            "align": "wide"
+            "imageCount": 2,
+            "allowResize": false,
+            "isListItem": true
         },
-        "innerBlocks": [],
+        "innerBlocks": [
+            {
+                "clientId": "_clientId_0",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
+                    "alt": "title",
+                    "caption": [],
+                    "id": 1,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            },
+            {
+                "clientId": "_clientId_1",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
+                    "alt": "title",
+                    "caption": [],
+                    "id": 2,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            }
+        ],
         "originalContent": "<ul class=\"wp-block-gallery columns-2 is-cropped alignwide\">\n\t<li class=\"blocks-gallery-item\">\n\t\t<figure>\n\t\t\t<img src=\"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==\" data-id=\"1\" alt=\"title\" class=\"wp-image-1\" />\n\t\t</figure>\n\t</li>\n\t<li class=\"blocks-gallery-item\">\n\t\t<figure>\n\t\t\t<img src=\"data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=\" data-id=\"2\" alt=\"title\" class=\"wp-image-2\" />\n\t\t</figure>\n\t</li>\n</ul>"
     }
 ]
