+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-2.serialized.html
@@ -1,3 +1,9 @@
-<!-- wp:gallery {"columns":2,"linkTo":"none","align":"wide"} -->
-<figure class="wp-block-gallery alignwide columns-2 is-cropped"><ul class="blocks-gallery-grid"></ul></figure>
+<!-- wp:gallery {"columns":2,"linkTo":"none","imageCount":2} -->
+<figure class="wp-block-gallery columns-2 is-cropped"><ul class="blocks-gallery-grid"><!-- wp:image {"id":1,"linkDestination":"none"} -->
+<figure class="wp-block-image"><img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==" alt="title" class="wp-image-1"/></figure>
+<!-- /wp:image -->
+
+<!-- wp:image {"id":2,"linkDestination":"none"} -->
+<figure class="wp-block-image"><img src="data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=" alt="title" class="wp-image-2"/></figure>
+<!-- /wp:image --></ul></figure>
 <!-- /wp:gallery -->
