+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-3.json
@@ -4,27 +4,42 @@
         "name": "core/gallery",
         "isValid": true,
         "attributes": {
-            "images": [
-                {
-                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
-                    "alt": "title",
-                    "caption": []
-                },
-                {
-                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
-                    "alt": "title",
-                    "caption": []
-                }
-            ],
-            "ids": [
-                null,
-                null
-            ],
             "imageCrop": true,
             "linkTo": "none",
-            "align": "wide"
+            "imageCount": 2,
+            "allowResize": false,
+            "isListItem": true
         },
-        "innerBlocks": [],
+        "innerBlocks": [
+            {
+                "clientId": "_clientId_0",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==",
+                    "alt": "title",
+                    "caption": [],
+                    "id": null,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            },
+            {
+                "clientId": "_clientId_1",
+                "name": "core/image",
+                "isValid": true,
+                "attributes": {
+                    "url": "data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=",
+                    "alt": "title",
+                    "caption": [],
+                    "id": null,
+                    "linkDestination": "none",
+                    "isListItem": false
+                },
+                "innerBlocks": []
+            }
+        ],
         "originalContent": "<ul class=\"wp-block-gallery columns-2 is-cropped alignwide\">\n\t<li class=\"blocks-gallery-item\">\n\t\t<figure>\n\t\t\t<img src=\"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAACklEQVR4nGMAAQAABQABDQottAAAAABJRU5ErkJggg==\" alt=\"title\" />\n\t\t</figure>\n\t</li>\n\t<li class=\"blocks-gallery-item\">\n\t\t<figure>\n\t\t\t<img src=\"data:image/jpeg;base64,/9j/2wBDAAMCAgICAgMCAgIDAwMDBAYEBAQEBAgGBgUGCQgKCgkICQkKDA8MCgsOCwkJDRENDg8QEBEQCgwSExIQEw8QEBD/yQALCAABAAEBAREA/8wABgAQEAX/2gAIAQEAAD8A0s8g/9k=\" alt=\"title\" />\n\t\t</figure>\n\t</li>\n</ul>"
     }
 ]
